+++ RecurrentFF/benchmarks/seq_mnist/seq.py
@@ -4,10 +4,11 @@ import torch
 from torch.utils.data import DataLoader, Dataset
 import wandb
 import numpy as np
+from RecurrentFF.model.data_scenario.static_single_class import SingleStaticClassTestData
 
 from RecurrentFF.model.model import RecurrentFFNet
 from RecurrentFF.settings import Settings
-from RecurrentFF.util import DataConfig, SingleStaticClassTestData, TrainInputData, TrainLabelData, set_logging
+from RecurrentFF.util import DataConfig, TrainInputData, TrainLabelData, set_logging
 
 INPUT_SIZE = 4
 NUM_CLASSES = 10
